install dependency `git` if not available
Fresh install of termux (at least in my case) does not include `git` which is a dependency required for cloning the Wyoming Satellite repo and the optional Wyoming OpenWakeWord repo if selected.

PR adds a check to see if `git` is installed and installs if if not.  Added code was copied from the equivalent check for wget.